Adding A Field To Taxonomy
This will need to be changed, but the code contained in this commit
demonstrates how to hard code a field to a taxonomy term display. This
will need to be changed to be more flexible, and starting with Drupal
8.8, a new set of methods will need to be used to replace the
entity_get_display and entity_get_form_display functions being used. See
https://www.drupal.org/node/2835616 and
https://drupal.stackexchange.com/questions/222260/add-content-type-field-programmatically
for more details. This relates to issue #3.