Finish Implementing TCBWebResource Logic In End Points
The logic added in TCBWebResource to simplify the end point code has now
been implemented across all end points. There is one exception in the
TCBUserResource class when searching on email, but that is a special
case that I am choosing not to account for in the TCBWebResource class
(for now). I also fixed a bug in TCBWebResource where getting a name
would return null because it was checking for a tid variable which only
exists when searching by tid, not name.