Embedded Entity Information
Information about embedded (referenced) entities within terms is now
also returned. For example, the default role on a user term is no longer
just the term ID of the term representing that role object. Instead, all
of the information that one would receive from querying the role
endpoint with that TID is included in the user JSON. This required some
heavy overhaul of currently existing field value extraction techniques.
Some cleanup may still be required in the aftermath of the overhaul, but
the bulk of it is done. This completes #11.